Fix warning with an explicit conversion

+++ QGLViewer/manipulatedFrame.cpp
@@ -219,7 +219,7 @@ use mouse speed. Currently used to trigger spinning in mouseReleaseEvent(). */
 void ManipulatedFrame::computeMouseSpeed(const QMouseEvent* const e)
 {
 	const QPoint delta = (e->pos() - prevPos_);
-	const qreal dist = sqrt(delta.x()*delta.x() + delta.y()*delta.y());
+	const qreal dist = sqrt(qreal(delta.x()*delta.x() + delta.y()*delta.y()));
 	delay_ = last_move_time.restart();
 	if (delay_ == 0)
 		// Less than a millisecond: assume delay = 1ms
